Fixed #73 errors with draft emails
Saving a draft would show the previous form values. Also a bad
recipient in a saved draft would prevent using that draft.

Moved code around so the $form email_form object contains the data the
request should work on. Before it was a mix of accesing the request,
$form or fetched database data.

Moved the mailto to two <select>s logic to class email_form. Fixes needing
$form to get mailto and needing mailto to create the form.